Add module building3d.geom.bboxes
