+++ building3d/geom/polygon/ispointinside.py
@@ -3,7 +3,7 @@ from numba import njit
 
 from building3d.config import GEOM_ATOL
 from building3d.geom.points import are_points_coplanar
-from building3d.geom.points import is_point_inside_bbox
+from building3d.geom.bboxes import is_point_inside_bbox
 from building3d.geom.points import is_point_on_segment
 from building3d.geom.points.distance import distance_point_to_edge
 from building3d.geom.polygon.edges import polygon_edges
