+++ building3d/geom/polygon/touching.py
@@ -1,7 +1,7 @@
 import numpy as np
 from numba import njit
 
-from building3d.geom.points import bounding_box
+from building3d.geom.bboxes import bounding_box
 from building3d.geom.points import new_point_between_2_points
 from building3d.geom.points import points_equal
 from building3d.geom.polygon.ispointinside import is_point_at_boundary
